Added new options for damage numbers

+++ Player/Overrides/PlayerStatsEx.cs
@@ -735,7 +735,7 @@ namespace ChampionsOfForest
 				amount = ModdedPlayer.instance.DealDamageToShield(-amount);
 				Health -= amount;
 				HealthTarget -= amount * 3;
-				Network.NetworkManager.SendPlayerHitmarker(transform.position, (int)amount);
+				Network.NetworkManager.SendPlayerHitmarker(transform.position, amount);
 			}
 			else
 			{
